// api/prisma/enums/product-ids.ts
/**
 * Products enum with hardcoded nanoid values. The nanoid values have a
 * length of 5 characters to prevent duplicates, since all nanoid's
 * generated by seeding are a length of 10.
 */
export enum ProductIds {
  tshirt = 'KIIOc',
  pants = '0we-a',
  shorts = 'Ge2N3',
  sweater = '88tOb',
  hat = 'RlV0I',
  hoodie = 'pHD3V',
}